refactor: create a get client auth util
