+++ app/(pages)/(home)/page.tsx
@@ -1,28 +1,14 @@
 import type { Metadata } from 'next';
-import { cookies } from 'next/headers';
 import Home from '@/_features/home/view';
-import { InternalApiFetcher } from '@/_shared/utils/fetcher';
-import type { AuthType } from '@/_shared/types/auth';
+import { getClientAuth } from '@/_shared/utils/get-client-auth';
 
 export const metadata: Metadata = {
   title: 'inLive Room',
   description: 'Conference room for real-time video and audio calls',
 };
 
-const getCurrentAuth = async () => {
-  const cookie = cookies().toString();
-  const currentAuth: AuthType.CurrentAuthInternalResponse =
-    await InternalApiFetcher.get('/api/auth/current', {
-      headers: {
-        cookie: cookie,
-      },
-    });
-
-  return currentAuth;
-};
-
 export default async function Page() {
-  const currentAuth = await getCurrentAuth();
+  const currentAuth = await getClientAuth();
   const currentUser = currentAuth.data ? currentAuth.data : undefined;
 
   return <Home currentUser={currentUser} />;
